[vuop_gen] move logic for viext
